+++ src/ska/__init__.py
@@ -1,5 +1,5 @@
 __title__ = 'ska'
-__version__ = '1.4'
+__version__ = '1.4.2'
 __build__ = 0x00000E
 __author__ = 'Artur Barseghyan'
 __copyright__ = 'Copyright (c) 2013 Artur Barseghyan'